requirements/isaac_requirements.txt
addict==2.4.0
aiodns==3.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.4
aiohttp-cors==0.7.0
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.6.2
asteval==0.9.32
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
av==15.0.0
blessed==1.21.0
cchardet==2.1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
coloredlogs==15.0.1
construct==2.10.68
cycler==0.11.0
decorator==5.2.1
diffusers==0.32.2
exceptiongroup==1.2.2
executing==2.2.0
expandvars==1.1.1
fastapi==0.110.0
ffmpeg==1.4
filelock==3.15.4
flatbuffers==25.2.10
frozenlist==1.4.1
fsspec==2024.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
gunicorn==22.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.1
httpx==0.25.2
humanfriendly==10.0
idna==3.6
imageio==2.22.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inputs==0.5
ipdb==0.13.13
ipython==8.20.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.4
lcm==1.5.0
lmdb==1.6.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
msgpack-numpy==0.4.8
multidict==6.0.5
nest-asyncio==1.5.6
networkx==3.4.2
numba==0.59.1
numpy==1.26.0
numpy-quaternion==2023.0.3
nvsmi==0.4.2
oauthlib==3.2.2
onnx==1.18.0
onnxruntime==1.19.2
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
packaging==23.0
params_proto==2.13.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
Pint==0.20.1
platformdirs==4.3.8
plotly==5.3.1
portalocker==2.7.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.0.9
pyperclip==1.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post1
qrcode==7.4.2
ray==2.47.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rich==14.0.0
rpds-py==0.26.0
rsl-rl-lib==2.2.1
s3transfer==0.6.1
scipy
selenium==4.14.0
sentry-sdk==2.33.0
setproctitle==1.3.6
setuptools-scm==8.0.4
shtab==1.7.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
termcolor==3.1.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tyro==0.9.26
urllib3==2.2.3
uvicorn==0.29.0
vuer==0.0.32rc7
wandb==0.20.0
watchdog==4.0.0
wcwidth==0.2.13
webbot==0.34
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.23.0